requirements.txt
streamlit
google-generativeai
pyttsx3
gTTS